Weather alerts/dashboard/requirements.txt
kafka-python==2.0.2
python-dotenv==1.0.0
requests==2.31.0
streamlit==1.24.1